(client): Finished styling
