+++ client/src/components/Profile/Profile.css
@@ -3,6 +3,6 @@
   flex-direction: row;
   width: 100%;
   height: 15rem;
-  margin-top: 5rem;
+  margin-top: 3rem;
   justify-content: space-between;
 }
